chore(suite): rename index files
