+++ packages/suite-web/src/Main.tsx
@@ -7,9 +7,8 @@ import { initStore } from 'src/reducers/store';
 import { preloadStore } from 'src/support/suite/preloadStore';
 import { SENTRY_CONFIG } from '@suite-common/sentry';
 
-import Metadata from 'src/components/suite/Metadata';
 import Preloader from 'src/components/suite/Preloader';
-import { ToastContainer } from 'src/components/suite';
+import { Metadata, ToastContainer } from 'src/components/suite';
 import { ConnectedIntlProvider } from 'src/support/suite/ConnectedIntlProvider';
 import Resize from 'src/support/suite/Resize';
 import Protocol from 'src/support/suite/Protocol';
