+++ packages/suite/src/components/suite/Loading.tsx
@@ -14,10 +14,8 @@ type LoadingProps = {
     className?: string;
 };
 
-const Loading = ({ className }: LoadingProps) => (
+export const Loading = ({ className }: LoadingProps) => (
     <LoaderWrapper data-test="@suite/loading" className={className}>
         <Image width={80} height={80} image="SPINNER" />
     </LoaderWrapper>
 );
-
-export default Loading;
