+++ packages/suite/src/components/suite/SuiteLayout/MenuSecondary.tsx
@@ -21,10 +21,8 @@ interface MenuSecondaryProps {
     children: ReactNode;
 }
 
-const MenuSecondary = ({ children }: MenuSecondaryProps) => (
+export const MenuSecondary = ({ children }: MenuSecondaryProps) => (
     <AbsoluteWrapper>
         <Wrapper>{children}</Wrapper>
     </AbsoluteWrapper>
 );
-
-export default MenuSecondary;
