+++ packages/suite/src/components/suite/index.tsx
@@ -9,14 +9,14 @@ import PinInput from './PinInput';
 import NoRatesTooltip from './NoRatesTooltip';
 import { WordInput } from './WordInput';
 import { WordInputAdvanced } from './WordInputAdvanced';
-import Loading from './Loading';
-import BundleLoader from './BundleLoader';
+import { Loading } from './Loading';
+import { BundleLoader } from './BundleLoader';
 import { FiatValue } from './FiatValue';
 import { WebUsbButton } from './WebUsbButton';
 import { HiddenPlaceholder } from './HiddenPlaceholder';
 import { QuestionTooltip } from './QuestionTooltip';
 import TransactionsGraph from './TransactionsGraph';
-import DeviceInvalidModeLayout from './DeviceInvalidModeLayout';
+import { DeviceInvalidModeLayout } from './DeviceInvalidModeLayout';
 import { AppNavigationPanel } from './AppNavigationPanel';
 import { AppNavigationTooltip } from './AppNavigationTooltip';
 import { AppNavigation } from './AppNavigation';
@@ -26,18 +26,19 @@ import { AccountLabeling, AddressLabeling, WalletLabeling, MetadataLabeling } fr
 import { FormattedCryptoAmount } from './FormattedCryptoAmount';
 import { FormattedNftAmount } from './FormattedNftAmount';
 import { Sign } from './Sign';
-import AddAccountButton from './AddAccountButton';
+import { AddAccountButton } from './AddAccountButton';
 import { ToastContainer } from './ToastContainer';
 import TrezorLink from './TrezorLink';
 import { ReadMoreLink } from './ReadMoreLink';
 import { Modal, ModalProps } from './Modal';
 import { SkeletonRectangle, SkeletonCircle } from './Skeleton';
 import Notifications from './Notifications';
-import FormattedDate, { FormattedDateWithBullet } from './FormattedDate';
-import Metadata from './Metadata';
+import { FormattedDate } from './FormattedDate';
+import { FormattedDateWithBullet } from './FormattedDateWithBullet';
+import { Metadata } from './Metadata';
 import { HomescreenGallery } from './HomescreenGallery';
-import DeviceMatrixExplanation from './DeviceMatrixExplanation';
-import AccountFormCloseButton from './AccountFormCloseButton';
+import { DeviceMatrixExplanation } from './DeviceMatrixExplanation';
+import { AccountFormCloseButton } from './AccountFormCloseButton';
 import { TroubleshootingTips } from './TroubleshootingTips';
 import { ConnectDevicePrompt } from './ConnectDevicePrompt';
 import { DeviceButton } from './DeviceButton';
@@ -45,7 +46,7 @@ import { Coin } from './Coin';
 import { CoinsGroup } from './CoinsGroup/CoinsGroup';
 import { CoinsList } from './CoinsGroup/CoinsList';
 import CharacterCount from './CharacterCount';
-import CloseButton from './CloseButton';
+import { CloseButton } from './CloseButton';
 import TooltipSymbol from './TooltipSymbol';
 import StatusLight from './StatusLight';
 import { AmountUnitSwitchWrapper } from './AmountUnitSwitchWrapper';
